+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/MethodScope.java
@@ -21,7 +21,6 @@ import com.fasterxml.classmate.ResolvedTypeWithMembers;
 import com.fasterxml.classmate.members.ResolvedMethod;
 import com.github.victools.jsonschema.generator.impl.LazyValue;
 import java.lang.annotation.Annotation;
-import java.lang.reflect.AnnotatedParameterizedType;
 import java.lang.reflect.AnnotatedType;
 import java.lang.reflect.Method;
 import java.util.List;
@@ -44,7 +43,7 @@ public class MethodScope extends MemberScope<ResolvedMethod, Method> {
      * @param context the overall type resolution context
      */
     protected MethodScope(ResolvedMethod method, ResolvedTypeWithMembers declaringTypeMembers, TypeContext context) {
-        this(method, null, null, declaringTypeMembers, false, context);
+        this(method, null, null, declaringTypeMembers, null, context);
     }
 
     /**
@@ -54,24 +53,24 @@ public class MethodScope extends MemberScope<ResolvedMethod, Method> {
      * @param overriddenType alternative type for this method's return value
      * @param overriddenName alternative name for this method
      * @param declaringTypeMembers collection of the declaring type's fields and (other) methods
-     * @param fakeContainerItemScope whether this field/method scope represents only the container item type of the actual field/method
+     * @param fakeContainerItemIndex index of the container item on the generic field/method scope's declared type (e.g., in case of a List, it is 0)
      * @param context the overall type resolution context
      */
     protected MethodScope(ResolvedMethod method, ResolvedType overriddenType, String overriddenName,
-            ResolvedTypeWithMembers declaringTypeMembers, boolean fakeContainerItemScope, TypeContext context) {
-        super(method, overriddenType, overriddenName, declaringTypeMembers, fakeContainerItemScope, context);
+            ResolvedTypeWithMembers declaringTypeMembers, Integer fakeContainerItemIndex, TypeContext context) {
+        super(method, overriddenType, overriddenName, declaringTypeMembers, fakeContainerItemIndex, context);
     }
 
     @Override
     public MethodScope withOverriddenType(ResolvedType overriddenType) {
         return new MethodScope(this.getMember(), overriddenType, this.getOverriddenName(), this.getDeclaringTypeMembers(),
-                this.isFakeContainerItemScope(), this.getContext());
+                this.getFakeContainerItemIndex(), this.getContext());
     }
 
     @Override
     public MethodScope withOverriddenName(String overriddenName) {
         return new MethodScope(this.getMember(), this.getOverriddenType(), overriddenName, this.getDeclaringTypeMembers(),
-                this.isFakeContainerItemScope(), this.getContext());
+                this.getFakeContainerItemIndex(), this.getContext());
     }
 
     @Override
@@ -180,13 +179,7 @@ public class MethodScope extends MemberScope<ResolvedMethod, Method> {
     @Override
     public <A extends Annotation> A getContainerItemAnnotation(Class<A> annotationClass) {
         AnnotatedType annotatedReturnType = this.getRawMember().getAnnotatedReturnType();
-        if (annotatedReturnType instanceof AnnotatedParameterizedType) {
-            AnnotatedType[] typeArguments = ((AnnotatedParameterizedType) annotatedReturnType).getAnnotatedActualTypeArguments();
-            if (typeArguments.length == 1) {
-                return typeArguments[0].getAnnotation(annotationClass);
-            }
-        }
-        return null;
+        return this.getContext().getTypeParameterAnnotation(annotationClass, annotatedReturnType, this.getFakeContainerItemIndex());
     }
 
     @Override